+++ NEWS.md
@@ -15,6 +15,7 @@
 * Implement mTLS Client-Side Authentication for FSSP client (APPPOCTOOL-62)
 * Add support for custom Keycloak base URL for JWKS endpoint ((MODSIDECAR-148)[https://folio-org.atlassian.net/browse/MODSIDECAR-148])
 * Upgrade Keycloak to 26.4.6 (KEYCLOAK-73)
+* Register JacksonProvider in ResteasyClient Keycloak for TLS connections (KEYCLOAK-90)
 
 ### Migration:
 * Replace KeycloakSecretUtils with SecureStoreKeyProvider, add `application.secret-store.environment=${SECURE_STORE_ENV:folio}` to application.yaml (for Ramsons and Sunflower: add `application.secret-store.environment=${SECURE_STORE_ENV:${ENV:folio}}` to application.yaml).